+++ transport/tchannel/handler.go
@@ -188,12 +188,22 @@ func (h handler) callHandler(ctx context.Context, call inboundCall, responseWrit
 		ctx = tchannel.ExtractInboundSpan(ctx, tcall.InboundCall, headers.Items(), tracer)
 	}
 
+	buf := bufferpool.Get()
+	defer bufferpool.Put(buf)
+
 	body, err := call.Arg3Reader()
 	if err != nil {
 		return err
 	}
-	defer body.Close()
-	treq.Body = body
+
+	if _, err = buf.ReadFrom(body); err != nil {
+		return err
+	}
+	if err = body.Close(); err != nil {
+		return err
+	}
+
+	treq.Body = buf
 
 	if err := transport.ValidateRequest(treq); err != nil {
 		return err