+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- observability: Add request and response payload size histogram
+### Fixed
 - yarpc: service name can contain `_` now.
 
 ## [1.48.0] - 2020-10-07